refactor(webhook) : followed kebab conventions for folders and files
