+++ src/constants/verification-codes.js
