+++ src/controllers/msg91/sent-mobile-otp.controller.js
@@ -50,17 +50,17 @@
 // start of the sent Mobile OTP Controller
 
 import _ from "lodash";
-import statusCodes from "../../constants/httpStatusCodes.js";
-import catchAsync from "../../utils/catchAsync.js";
+import statusCodes from "../../constants/status-codes.js";
+import catchAsync from "../../utils/catch-async.js";
 import { failed_response, success_response } from "../../utils/response.js";
-import AppError from "../../utils/AppError.js";
-import AppErrorCode from "../../constants/AppErrorCodes.js";
-import { FirebaseCheckPhoneExistOrNot } from "../../firebase/firebaseUtils.js";
-import { generateOTP, generateReferenceID } from "../../utils/generateIds.js";
-import { verificationCodes } from "../../constants/verificationCode.js";
-import { fiveMinutesFromNow } from "../../utils/dateHandlers.js";
-import OTP from "../../models/Otp.model.js";
-import sentVerifyMobileOTP from "./sentOtpToMobile.js";
+import AppError from "../../utils/app-error.js";
+import AppErrorCode from "../../constants/app-error-codes.js";
+import { FirebaseCheckPhoneExistOrNot } from "../../firebase/firebase-utils.js";
+import { generateOTP, generateReferenceID } from "../../utils/generate-ids.js";
+import { verificationCodes } from "../../constants/verification-codes.js";
+import { fiveMinutesFromNow } from "../../utils/date-handlers.js";
+import OTP from "../../models/otp.model.js";
+import sentVerifyMobileOTP from "./sent-otp-to-mobile.js";
 import { parsePhoneNumberFromString } from 'libphonenumber-js'
 
 export const validatePhoneNumber = (phone) => {
