+++ src/controllers/zepto/sent-email-otp.js
@@ -12,8 +12,8 @@
  */
 import { SendMailClient } from "zeptomail";
 import env from "../../constants/env.js";
-import AppError from "../../utils/AppError.js";
-import statusCodes from "../../constants/httpStatusCodes.js";
+import AppError from "../../utils/app-error.js";
+import statusCodes from "../../constants/status-codes.js";
 
 const sentEmailOTP = async (data) => {
     try {
