+++ src/firebase/admin.js
@@ -1,5 +1,5 @@
 import admin from "firebase-admin";
-import serviceAccount from "./disputefirebaseKey.js";
+import serviceAccount from "./dispute-firebase-key.js";
 
 admin.initializeApp({
     credential: admin.credential.cert(serviceAccount)
