+++ src/models/dispute-log.model.js
@@ -48,7 +48,7 @@
  */
 
 import { Model, DataTypes } from 'sequelize';
-import sequelize from '../config/database.js';
+import sequelize from '../config/database.config.js';
 
 class DisputeLog extends Model {
     static associate(models) {
