+++ src/models/dispute.model.js
@@ -46,7 +46,7 @@
  */
 
 import { Model, DataTypes } from 'sequelize';
-import sequelize from '../config/database.js';
+import sequelize from '../config/database.config.js';
 
 class Dispute extends Model {
     static associate(models) {
