+++ src/models/gstin.model.js
@@ -37,7 +37,7 @@
  * }
  */
 import { Model, DataTypes } from 'sequelize';
-import sequelize from '../config/database.js';
+import sequelize from '../config/database.config.js';
 
 class GSTIN extends Model {
     // static associate(models) {}
