+++ src/models/merchant.model.js
@@ -60,7 +60,7 @@
  *   - hasMany DisputeLog (as: disputeLogs)
  */
 import { Model, DataTypes } from 'sequelize';
-import sequelize from '../config/database.js';
+import sequelize from '../config/database.config.js';
 
 class Merchant extends Model {
     static associate(models) {
