+++ src/models/notification.model.js
@@ -50,7 +50,7 @@
  * }
  */
 import { DataTypes, Model } from 'sequelize';
-import sequelize from '../config/database.js';
+import sequelize from '../config/database.config.js';
 
 class Notification extends Model {
     /**
