+++ src/models/staff-assign-state.model.js
@@ -29,7 +29,7 @@
  * @index merchant_id
  */
 import { Model, DataTypes } from 'sequelize';
-import sequelize from '../config/database.js';
+import sequelize from '../config/database.config.js';
 
 class StaffAssignmentState extends Model {
     // static associate(models) {}
