+++ src/models/staff.model.js
@@ -76,7 +76,7 @@
  * - first_name, last_name
  */
 import { Model, DataTypes } from 'sequelize';
-import sequelize from '../config/database.js';
+import sequelize from '../config/database.config.js';
 
 class Staff extends Model {
     static associate(models) {
