+++ src/models/user-role.model.js
@@ -43,7 +43,7 @@
  *   - hasMany: Staff (foreignKey: userRole, as: staffMembers)
  */
 import { Model, DataTypes } from 'sequelize';
-import sequelize from '../config/database.js';
+import sequelize from '../config/database.config.js';
 
 class UserRole extends Model {
     static associate(models) {
